refactored Wallet interface
